// src/model/enums/lookup-table-category.enum.ts
export enum LookupTableCategory {
	AVGEffectPrefab = "AVGEffectPrefab",
	TierMedalSprite = "TierMedalSprite",
	MainPageImage = "MainPageImage",
	StoreTagIcon = "StoreTagIcon",
	TierMedalPrefab = "TierMedalPrefab",
	RegionMap = "RegionMap",
	SpecialSignInPage = "SpecialSignInPage",
	CharacterPrefab = "CharacterPrefab",
	CharacterAsset = "CharacterAsset",
	CharacterMetagameAsset = "CharacterMetagameAsset",
	MonsterSkillIcon = "MonsterSkillIcon",
	EmotIcon = "EmotIcon",
	EncounterOptionIcon = "EncounterOptionIcon",
	AvgFlagUI = "AvgFlagUI",
	Monster_SpSkillIcon = "Monster_SpSkillIcon",
	ItemIconMid = "ItemIconMid",
	ItemIconSmall = "ItemIconSmall",
	CurrencyIcon = "CurrencyIcon",
	ItemAnimationPrefab = "ItemAnimationPrefab",
	BuffAsset = "BuffAsset",
	UIOffset_Character = "UIOffset_Character",
	VFXPrefab = "VFXPrefab",
	CharacterAutograph = "CharacterAutograph",
	CharacterImage_AVATAR = "CharacterImage_AVATAR",
	CharacterImage_MID = "CharacterImage_MID",
	CharacterImage_LARGE = "CharacterImage_LARGE",
	CharacterPortrait_LARGE = "CharacterPortrait_LARGE",
	CharacterPortrait = "CharacterPortrait",
	CharacterImage_SD = "CharacterImage_SD",
	CharacterTotem_MID = "CharacterTotem_MID",
	MetagameSpine = "MetagameSpine",
	CutsceneSequence = "CutsceneSequence",
	CutsceneActor = "CutsceneActor",
	LevelData_LevelScenePrefab = "LevelData_LevelScenePrefab",
	CutsceneDialogAsset = "CutsceneDialogAsset",
	Audio_Music = "Audio_Music",
	Audio_SFX = "Audio_SFX",
	LevelData_Actor = "LevelData_Actor",
	LevelData_Event = "LevelData_Event",
	LevelData_WaveSet = "LevelData_WaveSet",
	LevelData_EnemySet = "LevelData_EnemySet",
	EncounterObject = "EncounterObject",
	UIOffset_Monster = "UIOffset_Monster",
	MonsterImage_SD = "MonsterImage_SD",
	VideoObject = "VideoObject",
	BattleField_SD = "BattleField_SD",
	MetagameSequence = "MetagameSequence",
	MetagameSequenceAsset = "MetagameSequenceAsset",
	TutorialArchive = "TutorialArchive",
	DynamicView = "DynamicView",
	Scene = "Scene",
	SpriteAtlas = "SpriteAtlas",
}